Antonio Gallardo's formatting and import cleanup patch.
I may have missed a few things in the patch.  Also, I could not update to the new version of util concurrent in the project.xml, since maven couldn't find the 1.3.4 specified in the patch.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/jcs/trunk@224095 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/jcs/auxiliary/remote/server/RemoteCacheServerInfo.java
@@ -40,6 +40,6 @@ public class RemoteCacheServerInfo
      */
     protected static VMID vmid = new VMID();
     /** Description of the Field */
-    public static long listenerId = ( long ) vmid.hashCode();
+    public static long listenerId = vmid.hashCode();
 
 }
